+++ src/main/java/server/dtobject/segment/CreateSegment.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record CreateSegment ( Long pdi_inicial,
                              Long pdi_final,
-                             Float distancia,
+                             Double distancia,
                              String aviso,
                              Boolean acessivel){
 }
